% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{actifUser}
\alias{actifUser}
\title{actifUser}
\usage{
actifUser(conn, idSession)
}
\arguments{
\item{idSession}{identifiant de la session de l'utilisateur}
}
\value{
True si l'utilisateur est actif et False sinon
}
\description{
Permet de verifier que l'utilisateur est bien actif et donc autorise a exporter des donnees de la base.
Elle fait appel a la fonction verifIdSession(...), et verifie dans la table des usager, si l'utilisateur est actif.
}

