% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{activeSession}
\alias{activeSession}
\title{activeSession}
\usage{
activeSession(conn, idProject, idSession, idResult)
}
\arguments{
\item{idProject}{identifiant du projet}

\item{idSession}{identifiant de la session de l'utilisateur}
}
\value{
TRUE if session active ou FALSE si session inactive
}
\description{
Verifie que la session de l'utilisateur est bien active.
Elle fait appel a la fonction getEndSessionDate(...) pour la comparaison avec la date courante.
}

