% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{buildColonneFait}
\alias{buildColonneFait}
\title{buildColonneFait}
\usage{
buildColonneFait(con, idProject, idSession, colonneFait)
}
\arguments{
\item{idProject}{Identifiant du projet}

\item{idSession}{Identifiant de la session de l'utilisateur}

\item{colonneFait}{Liste des champs a afficher}
}
\value{
La liste des champs a afficher
}
\description{
Fonction qui construit les colonnes a afficher lors de l'export des faits
}

