% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{buildColonneVisite}
\alias{buildColonneVisite}
\title{buildColonneVisite}
\usage{
buildColonneVisite(con, idProject, idSession, colonneVisite)
}
\arguments{
\item{idProject}{identifiant du projet}

\item{idSession}{identifiant de la session de l'utilisateur}

\item{colonneVisite}{Liste des champs a afficher}
}
\value{
La liste des champs a afficher
}
\description{
Fonction qui construit les colonnes a afficher lors de l'export de la liste des visites des patients
}

