% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{checkRole}
\alias{checkRole}
\title{checkRole}
\usage{
checkRole(conn, idProject, idSession)
}
\arguments{
\item{idProject}{Identifiant du projet}

\item{idSession}{Identifiant de la session de l'utilisateur}
}
\value{
0 lorsque l'utilisateur n'est pas autorisé a visualiser l'ensemble des champs et 1 sinon.
}
\description{
Fonction qui determine l'autorisation dont jouit l'utilisateur.                  
De cette autorisation dependra la liste des colonnes a afficher au moment de l'export des donnees.
}

