% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dbConnexion}
\alias{dbConnexion}
\title{Fonction de connexion a la base de donnees}
\usage{
dbConnexion(connexionFile)
}
\arguments{
\item{connexionInfo}{datatable contenant tous les informations de connexion}
}
\description{
Fonction de connexion a la base de donnees
}

