% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{factCollection}
\alias{factCollection}
\title{factCollection}
\usage{
factCollection(idResult, idProject, idSession, connexionFile)
}
\arguments{
\item{idResult}{Numero du resultat}

\item{idProject}{Identifiant du projet}

\item{idSession}{Identifiant de la session}

\item{connexionFile}{le chemin du fichier de connexion}
}
\value{
La liste  des sejours correspondant
}
\description{
Cette fonction interroge la base de données et retourne la liste des faits realisee durant le sejour du patient.
}

