% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getEndSessionDate}
\alias{getEndSessionDate}
\title{getEndSessionDate}
\usage{
getEndSessionDate(conn, idProject, idSession, idResult)
}
\arguments{
\item{idProject}{identifiant du projet}

\item{idSession}{identifiant de la session de l'utilisateur}
}
\value{
renvoie la date de la fin de session de l'utilisateur
}
\description{
Permet de recuperer la date de la fin de session de l'utilisateur courant.
Elle fait appel aux differentes fonctions de verification : verifIdSession(...), verifIdProject(...), verifIdResult(...)
}

