% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getIdQuery}
\alias{getIdQuery}
\title{Fonction qui retourne le numero de la requete faite par l'utilisateur.}
\usage{
getIdQuery(conn, idResult)
}
\arguments{
\item{con}{Informations de connexion a la base de donnees}
}
\value{
idQuery
}
\description{
Fonction qui retourne le numero de la requete faite par l'utilisateur.
}

