% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getIdUser}
\alias{getIdUser}
\title{getIdUser}
\usage{
getIdUser(conn, idSession)
}
\arguments{
\item{conn}{Informations de connexion a la base de donnees}

\item{idProject}{Identifiant du projet}
}
\value{
user_id Identifiant de l'utilisateur
}
\description{
Fonction qui retourne le user_id
}

