% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getRoleUser}
\alias{getRoleUser}
\title{Verifie les droits de l'utilisateur}
\usage{
getRoleUser(conn, idProject, idSession)
}
\arguments{
\item{idProject}{identifiant du projet}

\item{idSession}{identifiant de la session de l'utilisateur}

\item{con}{Informations de connexion a la base de donnees}
}
\description{
Verifie les droits de l'utilisateur
}

