% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getValueFromConfFile}
\alias{getValueFromConfFile}
\title{getValueFromConfFile}
\usage{
getValueFromConfFile(file, pattern)
}
\arguments{
\item{file}{le chemin du fichier a lire}

\item{pattern}{le champ a lire}
}
\description{
Fonction pour la lecture de fichier de connexion
}

