% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{package}
\name{i2b2ExportData}
\alias{i2b2ExportData}
\alias{i2b2ExportData-package}
\title{i2b2ExportData}
\description{
Dans ce package, vous trouverez des fonctions de connexion a la base de donnees, de requetage et de restitution de resultats de requete
}

