% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{patientCollection}
\alias{patientCollection}
\title{patientCollection}
\usage{
patientCollection(idResult, idProject, idSession, connexionFile)
}
\arguments{
\item{idResult}{Numero du resultat}

\item{idProject}{Identifiant du projet}

\item{idSession}{Identifiant de la session}

\item{connexionFile}{le chemin du fichier de connexion}
}
\value{
La liste des sejours correspondant
}
\description{
A partir du numero du resultat de la requete d'i2b2, cette fonction renvoie la liste des patients correspondant.
}

