% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{verifIdProject}
\alias{verifIdProject}
\title{verifIdProject}
\usage{
verifIdProject(conn, idProject)
}
\arguments{
\item{conn}{information de connexion}

\item{idProjet}{L'idProjet a verifier}
}
\value{
True si l'idProjet existe
}
\description{
Verifie si l'identifiant du projet est valide.
}

