% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{verifIdResult}
\alias{verifIdResult}
\title{verifIdResult}
\usage{
verifIdResult(conn, idResult)
}
\arguments{
\item{conn}{information de connexion}

\item{idResult}{L'idResult a verifier}
}
\value{
True si l'idResult existe
}
\description{
Verifie si l'identifiant du projet est valide.
}

