% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{verifIdSession}
\alias{verifIdSession}
\title{verifIdSession}
\usage{
verifIdSession(conn, idSession)
}
\arguments{
\item{conn}{Information de connexion}

\item{idSession}{L'idSession a verifier}
}
\value{
True si l'idSession existe
}
\description{
Verifie si l'identifiant du projet est valide.
}

