% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{verifQueryOwner}
\alias{verifQueryOwner}
\title{verifQueryOwner}
\usage{
verifQueryOwner(conn, idSession, idProject, idResult)
}
\arguments{
\item{idSession}{Identifiant de la session de l'utilisateur}

\item{idProject}{Identifiant du projet}

\item{idResult}{Numéro du résultat}
}
\value{
True si l'utilisateur est le proprietaire de la requete
}
\description{
Cette fonction permet de verifier que l'utilisateur courant est bien l'auteur de la requete.
Avant cette verification, cette fonction fait appel aux differentes fonctions de verification : verifIdSession(...), verifIdProject(...). verifIdResult(...)
}

