% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{visitCollection}
\alias{visitCollection}
\title{visitCollection}
\usage{
visitCollection(idResult, idProject, idSession, connexionFile)
}
\arguments{
\item{idResult}{Numero du resultat}

\item{idProject}{Ndentifiant du projet}

\item{idSession}{Identifiant de la session}
}
\value{
La liste  des sejours correspondant
}
\description{
A partir du numero du resultat de la requete d'i2b2, cette fonction renvoie la liste des visites correspondant.
}

