% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{buildColonnePatient}
\alias{buildColonnePatient}
\title{buildColonnePatient}
\usage{
buildColonnePatient(con, idProject, idSession, colonnePatient)
}
\arguments{
\item{idProject}{Identifiant du projet}

\item{idSession}{Identifiant de la session de l'utilisateur}

\item{colonnePatient}{Liste des champs a afficher}
}
\value{
La liste des champs a afficher
}
\description{
Fonction qui construit les colonnes a afficher lors de l'export des patients
}

