% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getRoleUser}
\alias{getRoleUser}
\title{getRoleUser}
\usage{
getRoleUser(conn, idProject, idSession)
}
\arguments{
\item{conn}{Informations de connexion a la base de donnees}

\item{idProject}{Identifiant du projet}

\item{idSession}{Identifiant de la session de l'utilisateur}
}
\value{
le droit de l'utilisateur en terme de visualisation de donnees
}
\description{
Verifie les droits de l'utilisateur
}

