% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{package}
\name{i2b2ExportData}
\alias{i2b2ExportData}
\alias{i2b2ExportData-package}
\title{i2b2ExportData}
\description{
Dans ce package, vous trouverez des fonctions pour l'export des données des requêtes
}

